rm history module
