+++ modules/pricefeed/handle_periodic_operations.go
@@ -2,6 +2,7 @@ package pricefeed
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/go-co-op/gocron"
 	"github.com/rs/zerolog/log"
@@ -19,6 +20,13 @@ func (m *Module) RegisterPeriodicOperations(scheduler *gocron.Scheduler) error {
 		utils.WatchMethod(m.updatePrice)
 	}); err != nil {
 		return fmt.Errorf("error while setting up pricefeed period operations: %s", err)
+	}
+
+	// Fetch total supply of tokens every 1hr to store historical price data
+	if _, err := scheduler.Every(1).Hour().Do(func() {
+		utils.WatchMethod(m.updatePricesHistory)
+	}); err != nil {
+		return fmt.Errorf("error while setting up history period operations: %s", err)
 	}
 
 	return nil
@@ -55,5 +63,37 @@ func (m *Module) updatePrice() error {
 	}
 
 	return nil
+
+}
+
+// updatePricesHistory fetches total amount of coins in the system from RPC
+// and stores historical perice data inside the database
+func (m *Module) updatePricesHistory() error {
+	log.Debug().
+		Str("module", "pricefeed").
+		Str("operation", "pricefeed").
+		Msg("getting token price and market cap history")
+
+	// Get the list of tokens price id
+	ids, err := m.db.GetTokensPriceID()
+	if err != nil {
+		return fmt.Errorf("error while getting tokens price id: %s", err)
+	}
+
+	if len(ids) == 0 {
+		log.Debug().Str("module", "pricefeed").Msg("no traded tokens price id found")
+		return nil
+	}
+
+	// Get the tokens prices
+	prices, err := coingecko.GetTokensPrices(ids)
+	if err != nil {
+		return fmt.Errorf("error while getting tokens prices history: %s", err)
+	}
+	for _, price := range prices {
+		price.Timestamp = time.Now()
+	}
+
+	return m.db.SaveTokenPricesHistory(prices)
 
 }
